+++ HISTORY.md
@@ -11,6 +11,7 @@ V2.3.0 ()
 * Traverse devices' data when app watches input
 * Added snapshot in cov class
 * When ttl is less than 1 second then disable
+* Using set_online_check_host instead of set_online_check_ip
 
 
 V2.2.4 (2023-10-11)